backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
joblib
scikit-learn
pandas
numpy
requests
psycopg2-binary
xgboost
python-dotenv
tensorflow